requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9.7
Pillow>=10.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0 